Fix sequelize errors when using postgres dialect
This came down to an issue with the pg's library hybrid approach to
types. If you pass a callback, it uses that. If you don't, a promise is
returned.

Sequelize was hitting an error in our handling of the promise usecase.

Since it's complex to attempt to replicate the exact behaviour of the
API surface, I've changed the pg patch to instead rely on internal
events in the Query class.

This simplifies the patch a fair bit, but also means we have to obtain a
reference to the client through another means.