Fix async hook propagation in Next.js dev server
This one is a bit of a doozy for how short the fix ended up being. In
essence, right before Next starts serving a request, they away a
devReady promise to make sure the server is ready.

This devReady promise is shared across the dev server, so all requests
await the same promise before beginning.

However, it turns out that awaiting a promise creating a new async
resource with the trigger id set to that promise. This is normally fine,
but in this case it totally breaks our ability to track what happens
inside of the request handler.

This is because the devReady is created once when the server is created,
rather than once for each request.

Therefore, the easiest solution is to wrap the devReady field with a
property that creates a new promise that resolves when devReady
resolves.

The promise that's returned will have the correct trigger id, since it
was created inside of the request's async resources. This keeps the same
behavior for Next.js as much as possible while maintaining async context
propagation.